updates to the home page
