+++ src/Images/BreakingForm/breakingFormExport.js
@@ -1,5 +1,5 @@
-export const beyondReach  = require('./jpgs/BEYOND REACH FROM SERIES YELLOW 2015.jpg').default
-// export const breakingFormCover = require('./jpgs/BOOK web.jpg').default
-export const holdOnToYourHat = require('./jpgs/HOLD ON TO YOUR HAT out-take 1.jpg').default
-export const supporting = require('./jpgs/SUPPORTING_FROM_SERIES_BLUE_2017.jpg').default
-export const tooFineALine = require('./jpgs/TOO FINE A LINE FROM W(H)ARMING 2018.jpg').default
+export const beyondReach  = require('./jpgs/BEYOND_REACH_FROM_SERIES_YELLOW_2015.jpg')
+// export const breakingFormCover = require('./jpgs/BOOK web.jpg')
+export const holdOnToYourHat = require('./jpgs/HOLD ON TO YOUR HAT out-take 1.jpg')
+export const supporting = require('./jpgs/SUPPORTING_FROM_SERIES_BLUE_2017.jpg')
+export const tooFineALine = require('./jpgs/TOO FINE A LINE FROM W(H)ARMING 2018.jpg')
