+++ src/Images/about/aboutExport.js
@@ -1,3 +1,3 @@
-export const portraitByCP = require('./jpgs/PortraitByClarePark.jpg').default
-export const headShot = require('./jpgs/debbie green by clare park.jpg').default
-export const film = require('./mp4/DGMovementFilm.mp4').default
+export const portraitByCP = require('./jpgs/PortraitByClarePark.jpg')
+export const headShot = require('./jpgs/debbie green by clare park.jpg')
+export const film = require('./mp4/DGMovementFilm.mp4')
