+++ src/Images/actorMovement/actorMovementExport.js
@@ -1,3 +1,3 @@
-export const actorMovementCover = require('./jpgs/5 Actor_Movement.jpg').default
-export const books = require('./jpgs/books.jpg').default
+export const actorMovementCover = require('./jpgs/5 Actor_Movement.jpg')
+export const books = require('./jpgs/books.jpg')
