+++ src/Images/contact/contactExport.js
@@ -1,1 +1,1 @@
-export const seeYouSoon = require('./jpgs/hope_to_see_you_again_soon_from_curtains_2020.jpg').default
+export const seeYouSoon = require('./jpgs/hope_to_see_you_again_soon_from_curtains_2020.jpg')
