+++ src/Images/work/workExport.js
@@ -1,6 +1,6 @@
-export const breakingForm = require('./BOOK web.jpg').default
-export const landscapes = require('./Yr3dw108.JPG').default
-export const dancer = require('./THE PIN FACTORY 2019.jpg').default
-export const actorMovement = require('./SYCHRONICITY FROM THE SERIES FRIENDSHIP 2016.jpg').default
-export const creativeDirector = require('./SUPPORTING FROM SERIES BLUE 2017.jpg').default
-export const theLyric = require(`./THE LYRIC I number one FROM THE SERIES ANNE'S POEMS.jpg`).default
+export const breakingForm = require('./BOOK web.jpg')
+export const landscapes = require('./Yr3dw108.JPG')
+export const dancer = require('./THE PIN FACTORY 2019.jpg')
+export const actorMovement = require('./SYCHRONICITY FROM THE SERIES FRIENDSHIP 2016.jpg')
+export const creativeDirector = require('./SUPPORTING FROM SERIES BLUE 2017.jpg')
+export const theLyric = require(`./THE LYRIC I number one FROM THE SERIES ANNE'S POEMS.jpg`)
