+++ src/Pages/TheLyric.jsx
@@ -7,7 +7,7 @@ const Lyric = () => {
       "https://res.cloudinary.com/dahzswwzk/image/upload/v1633184811/THE_LYRIC_I_number_one_FROM_THE_SERIES_ANNE_S_POEMS_1_h1gfhs.jpg",
     position: "0px",
   };
-  const leftContent = (
+  const rightContent = (
     <>
       <div>
         <h5>Performative photographic collaboration</h5>
@@ -16,21 +16,21 @@ const Lyric = () => {
       </div>
     </>
   );
-  const rightContent = (
+  const leftContent = (
     <>
-        <h1>The Lyric</h1>
-        <p>
-          The Lyric “I”, 2021, is a narrative photographic series by Clare and
-          I, that interprets/re-images Anne Gruenberg’s poetry. I perform ‘The
-          Dancer’ who portrays the desolation and reawakening from Anne’s mental
-          breakdown over a 12year period in secure institutions. For Clare and I
-          The Dancer is the expressive conduit to Anne’s work; Anne’s creativity
-          as a theatre designer was rekindled with the idea of The Dancer as
-          protagonist. The Dancer wears the ‘theatrical/archetypal’ red dress
-          Anne sourced; she moves in different landscapes with different props.
-          It is this combination that creates the heightened reality, symbolism
-          and spirituality redolent in each colour image.
-        </p>
+      <h1>The Lyric</h1>
+      <p>
+        The Lyric “I”, 2021, is a narrative photographic series by Clare and
+        I, that interprets/re-images Anne Gruenberg’s poetry. I perform ‘The
+        Dancer’ who portrays the desolation and reawakening from Anne’s mental
+        breakdown over a 12year period in secure institutions. For Clare and I
+        The Dancer is the expressive conduit to Anne’s work; Anne’s creativity
+        as a theatre designer was rekindled with the idea of The Dancer as
+        protagonist. The Dancer wears the ‘theatrical/archetypal’ red dress
+        Anne sourced; she moves in different landscapes with different props.
+        It is this combination that creates the heightened reality, symbolism
+        and spirituality redolent in each colour image.
+      </p>
     </>
   );
   const photoContent = {
@@ -55,7 +55,7 @@ const Lyric = () => {
       {
         title: "THE LYRIC I From The Series Anne's Poems 2021 © Clare Park",
         image:
-        "https://res.cloudinary.com/dahzswwzk/image/upload/v1622214651/THE_LYRIC_I_FROM_THE_SERIES_ANNE_S_POEMS_huhpml.jpg",
+          "https://res.cloudinary.com/dahzswwzk/image/upload/v1622214651/THE_LYRIC_I_FROM_THE_SERIES_ANNE_S_POEMS_huhpml.jpg",
       },
     ],
   };
@@ -65,8 +65,8 @@ const Lyric = () => {
       homeColour={"dark"}
       pageColour={"dark"}
       backgroundImage={backgroundImage}
-      leftContent={leftContent}
       rightContent={rightContent}
+      leftContent={leftContent}
       photoContent={photoContent}
       footerContent={null}
     />
